expose minimal required versions for ir and opset
Up to now, we set the opset import versions from the values
supported by the onnx library. This is an incorrect behavior.

We now store the maximum opset version required per domain,
based on the operators used in the graph. We then use the
onnx library to compute the minimal required IR version to
execute the model, based on these opset versions.

Fixes #22
Fixes #23